# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Call-a-Vet Gateway API
  version: 0.1.0
  description: >
    Public+private HTTP API for Call-a-Vet. Auth via Supabase JWT (Bearer).
    Primary login is Phone OTP (Supabase Auth: send code → verify code → session).
    RLS enforced at the DB layer. Idempotency on critical mutations via `Idempotency-Key`.
servers:
  - url: https://api.dev.callavet.mx
    description: Development
  - url: https://api.staging.callavet.mx
    description: Staging
  - url: https://api.callavet.mx
    description: Production
...
paths:
...
  /pets/{petId}/encounters:
    get:
      tags: [Notes & Care Plans]
      summary: List encounter timeline for a pet
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/petId" }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses: { "200": { $ref: "#/components/responses/ListEncounters" } }
  /encounters/{encounterId}:
    get:
      tags: [Notes & Care Plans]
      summary: Get full encounter timeline detail
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: encounterId
          required: true
          schema: { type: string, format: uuid }
      responses: { "200": { $ref: "#/components/responses/EncounterDetail" } }
...
  /sessions:
    get:
      tags: [Sessions]
      summary: List my sessions (user or vet)
      security: [{ bearerAuth: [] }]
      responses: { "200": { $ref: "#/components/responses/ListSessions" } }
...
  /sessions/{sessionId}/notes:
    get:
      tags: [Notes & Care Plans]
      summary: Get consultation notes for a session
      description: |
        Returns `{ data: [...] }` of notes.
        Access:
        - Owners: notes where the session's pet belongs to the user.
        - Vets: notes they authored, and all notes for sessions where they are the assigned vet.
        - Admins: all notes.
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: "#/components/parameters/sessionId" } ]
      responses: { "200": { $ref: "#/components/responses/ListNotes" } }
    post:
      tags: [Notes & Care Plans]
      summary: Upsert vet notes (SOAP/summary)
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: "#/components/parameters/sessionId" } ]
      requestBody: { $ref: "#/components/requestBodies/NotesUpsert" }
      responses: { "200": { $ref: "#/components/responses/Notes" } }
...
  /files/upload:
    post:
      tags: [Files]
      summary: Server-side upload to storage
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                path: { type: string }
                content: { type: string, description: base64 or data URL }
                contentType: { type: string }
                petId: { type: string, format: uuid }
                sessionId: { type: string, format: uuid }
                labels: { type: array, items: { type: string } }
                findings: { type: string }
                diagnosis_label: { type: string }
              required: [path, content]
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  path: { type: string }
                  contentType: { type: string }
  /files/download-url:
    get:
      tags: [Files]
      summary: Signed download URL
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: path
          required: true
          schema:
            type: string
      responses: { "200": { $ref: "#/components/responses/SignedUrl" } }
...
components:
  securitySchemes:
    adminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    petId: { in: path, name: petId, required: true, schema: { type: string, format: uuid } }
...
    sessionId: { in: path, name: sessionId, required: true, schema: { type: string, format: uuid } }
...
  requestBodies:
...
    NotesUpsert:
      content:
        application/json:
          schema:
            type: object
            properties:
              summary_text: { type: string }
              plan_summary: { type: string }
              assessment_text: { type: string }
              diagnosis_text: { type: string }
              follow_up_instructions: { type: string }
              next_follow_up_at: { type: string, format: date-time, nullable: true }
              severity: { type: string, enum: [low, medium, high, critical], nullable: true }
              pet_id: { type: string, format: uuid, nullable: true }
...
  responses:
...
    ListSessions:
      description: Sessions
      content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/Session" } } } } } }
...
    Notes:
      description: Vet notes
      content: { application/json: { schema: { $ref: "#/components/schemas/Notes" } } }
    ListNotes:
      description: Vet notes list
      content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/Notes" } } } } } }
...
    SignedUrl:
      description: Signed URL response
      content:
        application/json:
          schema:
            type: object
            properties:
              path: { type: string }
              url: { type: string, format: uri }
              token: { type: string, nullable: true }
              expires_in: { type: integer, nullable: true }
              expiresIn: { type: integer, nullable: true }
              method: { type: string }
              headers: { type: object, additionalProperties: true }
...
    ListEncounters:
      description: Encounter list
      content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/Encounter" } } } } } }
    EncounterDetail:
      description: Encounter timeline detail
      content:
        application/json:
          schema:
            type: object
            properties:
              encounter: { $ref: "#/components/schemas/Encounter" }
              notes: { type: array, items: { $ref: "#/components/schemas/Notes" } }
              imageCases: { type: array, items: { $ref: "#/components/schemas/ImageCase" } }
              carePlans: { type: array, items: { $ref: "#/components/schemas/CarePlan" } }
              files: { type: array, items: { $ref: "#/components/schemas/EncounterFile" } }
              appointment: { $ref: "#/components/schemas/Appointment", nullable: true }
              session: { $ref: "#/components/schemas/Session", nullable: true }
              healthProfile: { $ref: "#/components/schemas/PetHealthProfile", nullable: true }
...
  schemas:
...
    Session:
      type: object
      properties:
        id: { type: string, format: uuid }
        mode: { type: string, enum: [chat, video] }
        user_id: { type: string, format: uuid }
        vet_id: { type: string, format: uuid, nullable: true }
        pet_id: { type: string, format: uuid }
        status: { type: string }
        started_at: { type: string, format: date-time, nullable: true }
        ended_at: { type: string, format: date-time, nullable: true }
...
    Notes:
      type: object
      properties:
        id: { type: string, format: uuid }
        encounter_id: { type: string, format: uuid, nullable: true }
        session_id: { type: string, format: uuid }
        pet_id: { type: string, format: uuid }
        vet_id: { type: string, format: uuid, nullable: true }
        summary_text: { type: string, nullable: true }
        plan_summary: { type: string, nullable: true }
        assessment_text: { type: string, nullable: true }
        diagnosis_text: { type: string, nullable: true }
        follow_up_instructions: { type: string, nullable: true }
        next_follow_up_at: { type: string, format: date-time, nullable: true }
        severity: { type: string, enum: [low, medium, high, critical], nullable: true }
        created_at: { type: string, format: date-time }
    PetHealthProfile:
      type: object
      properties:
        pet_id: { type: string, format: uuid }
        allergies: { type: array, items: { type: string } }
        chronic_conditions: { type: array, items: { type: string } }
        current_medications:
          type: array
          items: { type: object, additionalProperties: true }
        vaccine_history:
          type: array
          items: { type: object, additionalProperties: true }
        injury_history:
          type: array
          items: { type: object, additionalProperties: true }
        procedure_history:
          type: array
          items: { type: object, additionalProperties: true }
        feed_profile: { type: object, additionalProperties: true }
        insurance: { type: object, additionalProperties: true }
        emergency_contacts:
          type: array
          items: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CarePlan:
      type: object
      properties:
        id: { type: string, format: uuid }
        encounter_id: { type: string, format: uuid, nullable: true }
        pet_id: { type: string, format: uuid }
        created_by_ai: { type: boolean }
        short_term: { type: string }
        mid_term: { type: string }
        long_term: { type: string }
...
    ImageCase:
      type: object
      properties:
        id: { type: string, format: uuid }
        encounter_id: { type: string, format: uuid, nullable: true }
        pet_id: { type: string, format: uuid }
        session_id: { type: string, format: uuid, nullable: true }
        image_url: { type: string, format: uri }
        labels: { type: array, items: { type: string } }
        findings: { type: string }
        diagnosis_label: { type: string }
    EncounterFile:
      type: object
      properties:
        id: { type: string, format: uuid }
        encounter_id: { type: string, format: uuid }
        pet_id: { type: string, format: uuid }
        session_id: { type: string, format: uuid, nullable: true }
        storage_path: { type: string }
        content_type: { type: string, nullable: true }
        labels: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
    Encounter:
      type: object
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid, nullable: true }
        appointment_id: { type: string, format: uuid, nullable: true }
        pet_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        vet_id: { type: string, format: uuid, nullable: true }
        video_room_id: { type: string, nullable: true }
        status: { type: string }
        started_at: { type: string, format: date-time, nullable: true }
        ended_at: { type: string, format: date-time, nullable: true }
        notes_count: { type: integer, nullable: true }
        image_cases_count: { type: integer, nullable: true }
        care_plans_count: { type: integer, nullable: true }
        files_count: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Appointment:
      type: object
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid, nullable: true }
        vet_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [scheduled, confirmed, active, completed, no_show, canceled] }
...